rename search filter utils module
