+++ fpan/views/search.py
@@ -13,7 +13,7 @@ from arches.app.views.search import get_nodegroups_by_datatype_and_perm
 from arches.app.views.search import get_permitted_nodegroups, select_geoms_for_results
 from arches.app.search.elasticsearch_dsl_builder import Bool, Match, Query, Nested, Term, Terms, GeoShape, Range, MinAgg, MaxAgg, RangeAgg, Aggregation, GeoHashGridAgg, GeoBoundsAgg, FiltersAgg, NestedAgg
 from fpan.utils.permission_backend import get_allowed_resource_ids
-from fpan.utils.filter import apply_advanced_docs_permissions
+from fpan.utils.search_filter import apply_advanced_docs_permissions
 
 class FPANSearchView(SearchView):
